refactor: reorganize components into sections directory
