+++ components/resume.js
@@ -1,18 +1,18 @@
 import { html } from '@rbardini/html'
 import colors from '../utils/colors.js'
-import Awards from './awards.js'
-import Certificates from './certificates.js'
-import Education from './education.js'
 import Header from './header.js'
-import Interests from './interests.js'
-import Languages from './languages.js'
 import Meta from './meta.js'
-import Projects from './projects.js'
-import Publications from './publications.js'
-import References from './references.js'
-import Skills from './skills.js'
-import Volunteer from './volunteer.js'
-import Work from './work.js'
+import Awards from './sections/awards.js'
+import Certificates from './sections/certificates.js'
+import Education from './sections/education.js'
+import Interests from './sections/interests.js'
+import Languages from './sections/languages.js'
+import Projects from './sections/projects.js'
+import Publications from './sections/publications.js'
+import References from './sections/references.js'
+import Skills from './sections/skills.js'
+import Volunteer from './sections/volunteer.js'
+import Work from './sections/work.js'
 
 /**
  * @param {import('../schema.d.ts').ResumeSchema} resume
