+++ components/sections/interests.js
@@ -1,8 +1,8 @@
 import { html } from '@rbardini/html'
-import Section from './general/section.js'
+import Section from '../general/section.js'
 
 /**
- * @param {import('../schema.d.ts').ResumeSchema['interests']} interests
+ * @param {import('../../schema.d.ts').ResumeSchema['interests']} interests
  * @returns {string | false}
  */
 export default function Interests(interests = []) {
