+++ components/sections/languages.js
@@ -1,8 +1,8 @@
 import { html } from '@rbardini/html'
-import Section from './general/section.js'
+import Section from '../general/section.js'
 
 /**
- * @param {import('../schema.d.ts').ResumeSchema['languages']} languages
+ * @param {import('../../schema.d.ts').ResumeSchema['languages']} languages
  * @returns {string | false}
  */
 export default function Languages(languages = []) {
