+++ components/sections/publications.js
@@ -1,12 +1,12 @@
 import { html } from '@rbardini/html'
-import markdown from '../utils/markdown.js'
-import Article from './general/article.js'
-import DateTime from './general/date-time.js'
-import Link from './general/link.js'
-import Section from './general/section.js'
+import markdown from '../../utils/markdown.js'
+import Article from '../general/article.js'
+import DateTime from '../general/date-time.js'
+import Link from '../general/link.js'
+import Section from '../general/section.js'
 
 /**
- * @param {import('../schema.d.ts').ResumeSchema['publications']} publications
+ * @param {import('../../schema.d.ts').ResumeSchema['publications']} publications
  * @returns {string | false}
  */
 export default function Publications(publications = []) {
