+++ components/sections/references.js
@@ -1,9 +1,9 @@
 import { html } from '@rbardini/html'
-import markdown from '../utils/markdown.js'
-import Section from './general/section.js'
+import markdown from '../../utils/markdown.js'
+import Section from '../general/section.js'
 
 /**
- * @param {import('../schema.d.ts').ResumeSchema['references']} references
+ * @param {import('../../schema.d.ts').ResumeSchema['references']} references
  * @returns {string | false}
  */
 export default function References(references = []) {
