+++ components/sections/skills.js
@@ -1,8 +1,8 @@
 import { html } from '@rbardini/html'
-import Section from './general/section.js'
+import Section from '../general/section.js'
 
 /**
- * @param {import('../schema.d.ts').ResumeSchema['skills']} skills
+ * @param {import('../../schema.d.ts').ResumeSchema['skills']} skills
  * @returns {string | false}
  */
 export default function Skills(skills = []) {
